feat(xlog)：Linux 的ConsoleLog 与其他平台对齐，从mars/comm 放到mars/xlog
