+++ mars/comm/unix/platform_comm.cc
@@ -51,36 +51,4 @@ int OSVerifyCertificate(const std::string& hostname, const std::vector<std::stri
 }
 
 }  // namespace comm
-
-#ifndef OHOS
-namespace xlog {
-void ConsoleLog(const XLoggerInfo* _info, const char* _log) {
-    if (NULL == _info || NULL == _log)
-        return;
-    static const char* levelStrings[] = {
-        "V",
-        "D",  // debug
-        "I",  // info
-        "W",  // warn
-        "E",  // error
-        "F"   // fatal
-    };
-    char strFuncName[128] = {0};
-    ExtractFunctionName(_info->func_name, strFuncName, sizeof(strFuncName));
-    const char* file_name = ExtractFileName(_info->filename);
-    char log[16 * 1024] = {0};
-    snprintf(log,
-             sizeof(log),
-             "[%s][%s][%s, %s, %d][%s\n",
-             levelStrings[_info->level],
-             NULL == _info->tag ? "" : _info->tag,
-             file_name,
-             strFuncName,
-             _info->line,
-             _log);
-    printf("%s", log);
-}
-}  // namespace xlog
-#endif  // ifndef OHOS
-
 }  // namespace mars
