+++ mars/xlog/CMakeLists.txt
@@ -62,6 +62,9 @@ elseif(ANDROID)
 elseif(OHOS)
     file(GLOB SELF_TEMP_SRC_FILES RELATIVE ${PROJECT_SOURCE_DIR} ohos/*.cc)
     list(APPEND SELF_SRC_FILES ${SELF_TEMP_SRC_FILES})
+elseif(UNIX)
+    file(GLOB SELF_TEMP_SRC_FILES RELATIVE ${PROJECT_SOURCE_DIR} unix/*.cc)
+    list(APPEND SELF_SRC_FILES ${SELF_TEMP_SRC_FILES})
 endif()
 
 # add_library(${PROJECT_NAME} STATIC ${SELF_SRC_FILES})
